refactor(clean-arch): extract UI orchestration behind adapters and shared flows
One dismiss flow (payload build + target-project resolution + reconcile tail)
replaces the three copies in App.jsx; /scan calls go through the api adapter;
the cancel-confirmation dialog spec moves out of useEvaluation into a view
module; useMapPageState splits into DOM/fetch/storage/pure-tree pieces; the
plugin-dimension cache and grade thresholds become store factories with
compatible default instances; App.jsx shrinks to composition (1523 -> 532).
